ref: Extracts Participant in a separate file.
